Move this file, to be maintained along with the other app code. Colab notebooks will need to reference app code.
